Documentation update; added a bunch of screenshots and fleshed out the README.markdown quite a lot. Approaching releasable state, fast.
Not sure if I should add a "table of contents", or if the document is Small Enough(TM) not to warrant one.

Also not sure if I like the mix of markdown and raw HTML, i.e. in the case of the definition list used for the config window description.