Fix GTK warnings by refactoring out deprecated things, and more.
This completely removes the use of Actions (used to be GtkAction, now
they live in GIO and are called GActions) since I found they added no
value. Also struggled semi-hard to find any documentation on how they
are supposed to be used, and failed in my attempts to use them. So, why
bother.

Switched the language standard to C99, which is a huge step up from not
specifying it all (yuck) as before. Not done a lot of code polishing
though, but it could certainly use some since it's not up to my current
standards.

Moved the "Showing 1/N" message in the Quick Open dialog to above the
list view, since living in the action button area was annoying to get
right with the new API (could have used introspection I just realized,
but this makes more sense and it's not as if we need to hella-conserve
vertical space).

Bumps version to 1.6.0, and updates copyright in at least the main
source file.