Port to the current plugin registration procedure.
Tihs was a bit sneaky, since the old way still worked, but there was
documentation at geany.org about how to port and get with the current
program so let's do that. Fewer globals, and more statics is always
welcome in my book, plus it feels more rigid and less magic now. Win!

Bumped the required API version to 225 from the ancient 147. Phew.